Moved auth api client into typescript-common.
