+++ deploy/helm/magda/charts/preview-map/templates/service.yaml
@@ -13,4 +13,3 @@ spec:
 {{- end }}
   selector:
     service: preview-map
-  type: "NodePort"
