+++ magda-authorization-api/package.json
@@ -14,7 +14,6 @@
     "@magda/typescript-common": "^0.0.29-SNAPSHOT",
     "body-parser": "^1.13.2",
     "express": "^4.13.1",
-    "isomorphic-fetch": "^2.2.1",
     "jsonwebtoken": "^7.4.1",
     "pg": "^6.4.0",
     "tsmonad": "^0.7.2",
