+++ magda-gateway/package.json
@@ -13,7 +13,7 @@
     "docker-build-prod": "create-docker-context-for-node-component --build --push --tag auto"
   },
   "dependencies": {
-    "@magda/authorization-api": "^0.0.29-SNAPSHOT",
+    "@magda/typescript-common": "^0.0.29-SNAPSHOT",
     "body-parser": "^1.13.2",
     "cheerio": "^0.22.0",
     "connect-ensure-login": "^0.1.1",
