+++ magda-gateway/src/createAuthRouter.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import ApiClient from "@magda/authorization-api/dist/ApiClient";
+import ApiClient from "@magda/typescript-common/dist/authorization-api/ApiClient";
 import Authenticator from "./Authenticator";
 import * as passport from "passport";
