+++ magda-typescript-common/package.json
@@ -21,6 +21,7 @@
     "@types/request": "^2.0.0",
     "@types/urijs": "^1.15.34",
     "@types/uuid": "^3.4.0",
+    "@magda/registry-aspects": "^0.0.29-SNAPSHOT",
     "chai": "^4.0.1",
     "jsverify": "^0.8.2",
     "lazy-seq": "^1.0.0",
@@ -29,7 +30,7 @@
     "uuid": "^3.1.0"
   },
   "dependencies": {
-    "@magda/registry-aspects": "^0.0.29-SNAPSHOT",
+    "isomorphic-fetch": "^2.2.1",
     "jsonwebtoken": "^7.4.1",
     "lodash": "^4.17.4",
     "moment": "^2.17.1",
