+++ magda-typescript-common/src/authorization-api/model.ts
